Fixes #3055 Disable certain tests when run under -race
