+++ auth/auth_test.go
@@ -14,7 +14,6 @@ import (
 	"errors"
 	"log"
 	"testing"
-	"time"
 
 	"github.com/couchbase/sync_gateway/base"
 	ch "github.com/couchbase/sync_gateway/channels"
@@ -111,27 +110,6 @@ func TestUserPasswords(t *testing.T) {
 	assert.False(t, user2.Authenticate("password"))
 	assert.True(t, user.Authenticate("letmein"))
 	assert.False(t, user.Authenticate("password"))
-}
-
-// Test that multiple authentications of the same user/password are fast.
-// This is an important check because the underlying bcrypt algorithm used to verify passwords
-// is _extremely_ slow (~100ms!) so we use a cache to speed it up (see password_hash.go).
-func TestAuthenticationSpeed(t *testing.T) {
-
-	gTestBucket := base.GetTestBucketOrPanic()
-	defer gTestBucket.Close()
-	auth := NewAuthenticator(gTestBucket.Bucket, nil)
-	user, _ := auth.NewUser("me", "goIsKewl", nil)
-	assert.True(t, user.Authenticate("goIsKewl"))
-
-	start := time.Now()
-	for i := 0; i < 1000; i++ {
-		assert.True(t, user.Authenticate("goIsKewl"))
-	}
-	durationPerAuth := time.Since(start) / 1000
-	if durationPerAuth > time.Millisecond {
-		t.Errorf("user.Authenticate is too slow: %v", durationPerAuth)
-	}
 }
 
 func TestSerializeUser(t *testing.T) {
